04/10

+++ api/app/Http/Controllers/CheckoutController.php
@@ -7,6 +7,7 @@ namespace OrchidEats\Http\Controllers;
  use OrchidEats\Http\Requests\OffPlatformOrderRequest;
  use OrchidEats\Http\Requests\SaveOrderRequest;
  use OrchidEats\Mail\NewOrder;
+ use OrchidEats\Mail\NonUserOrder;
  use Stripe\Stripe;
  use Stripe\Customer;
  use Stripe\Charge;
@@ -27,28 +28,37 @@ namespace OrchidEats\Http\Controllers;
 
          $chef = Chef::find($request->chef_id);
          $seller = User::find($chef->chefs_user_id);
+         $email = $seller->email;
+         $phone = $seller->profile->phone;
          $stripe_user_id = $seller->stripe_user_id;
          $order = $request->order;
          $order_total = 0;
          $deliveryFee = 0;
-         $serviceFee = 1.49;
-         $url = env('APP_URL') . '/upcoming-orders/' . $order['orders_user_id'];
+         $serviceFee = 0.99;
 
+//            Checks to see if there is an account
+        if ($order['orders_user_id']) {
+            $url = env('APP_URL') . '/upcoming-orders/' . $order['orders_user_id'];
+        }
             $customer = Customer::create(array(
                 'email' => $request->email,
                 'source' => $request->id
             ));
 
-         if ($order['order_details']['method'] == 'Pickup') {
-             $deliveryFee = 0;
-         } else if ($order['order_details']['method'] == 'Delivery') {
-             $deliveryFee = floatval($chef->delivery_fee);
-         }
+             if ($order['order_details']['method'] == 'Pickup') {
+                 $deliveryFee = 0;
+             } else if ($order['order_details']['method'] == 'Delivery') {
+                 $deliveryFee = floatval($chef->delivery_fee);
+             }
 
-            foreach($order['meal_details'] as $meal) {
-                    $price = Meal::find($meal['meal_id'])->price;
-                    $order_total = $order_total + ($price * $meal['quantity']);
-            }
+             if ($order['bundle'] == true ) {
+                 $order_total = $order['order_total'] - $serviceFee;
+             } else {
+                 foreach ($order['meal_details'] as $meal) {
+                     $price = Meal::find($meal['meal_id'])->price;
+                     $order_total = $order_total + ($price * $meal['quantity']);
+                 }
+             }
 
             if ($order_total > 0) {
                 $order_total += ($deliveryFee + $serviceFee);
@@ -56,7 +66,7 @@ namespace OrchidEats\Http\Controllers;
 
 //         creates fee for service
             $order_total = $order_total * 100;
-            $fee = ceil(($order_total * 0.03) + 179);
+            $fee = ceil(($order_total * 0.03) + 130);
 
             //TODO : fix the charge amounts for the fees
             $charge = Charge::create(array(
@@ -73,27 +83,36 @@ namespace OrchidEats\Http\Controllers;
             if ($charge) {
                 $saved = $chef->orders()
                     ->create(array(
-                        'orders_user_id' => $order['orders_user_id'],
-                        'orders_chef_id' => $chef->id,
+                        'orders_user_id' => $order['orders_user_id'] ?? null,
+                        'orders_chef_id' => $chef->chef_id,
                         'meal_details' => json_encode($order['meal_details']),
                         'customer_details' => json_encode($order['customer_details']),
                         'order_details' => json_encode($order['order_details']),
-                        'order_total' => (($order_total - $fee) / 100),
+                        'order_total' => (($order_total) / 100),
                         'payment_method' => $order['payment_method']
                     ));
 
                 if ($saved) {
-                    $user = User::find($order['orders_user_id']);
-                    $user->cart()->delete();
-                    $order['url'] = $url;
-                    $order['buyer'] = $user->first_name;
-                    \Mail::to($order['customer_details']['email'])->send(new NewOrder($order));
+                    $user = User::find($order['orders_user_id']) ?? null;
+                    if ($user != null) {
+                        $user->cart()->delete();
+                        $data['url'] = $url;
+                        $data['buyer'] = $user->first_name;
+                        \Mail::to($order['customer_details']['email'])->send(new NewOrder($data));
+                    } else {
+                        $data['order'] = $order['order_details'];
+                        $data['meals'] = $order['meal_details'];
+                        $data['total'] = (($order_total) / 100);
+                        $data['buyer'] = $order['customer_details']['name'];
+                        $data['email'] = $email;
+                        $data['phone'] = $phone;
+                        \Mail::to($order['customer_details']['email'])->send(new NonUserOrder($data));
+                    }
                 }
             }
 
          return response()->json([
              'status' => 'success',
-             'data' => [$order_total, $fee]
          ]);
         } catch (\Exception $ex) {
         return response()->json($ex->getMessage());
@@ -102,31 +121,44 @@ namespace OrchidEats\Http\Controllers;
 
     public function saveOrder (OffPlatformOrderRequest $request) {
         $order = $request;
-        $chef = Chef::find($order->chef_id);
+        $chef = Chef::find($order['chef_id']);
+        $chefUser = User::find($chef->chefs_user_id);
+        $email = $chefUser->email;
+        $phone = $chefUser->profile->phone;
         $order_total = 0;
         $deliveryFee = 0;
-        $serviceFee = 1.49;
-        $url = env('APP_URL') . '/upcoming-orders/' . $order['orders_user_id'];
+//        $serviceFee = 0.99;
 
+//        Checks to see if there is an account
+        if ($order['orders_user_id']) {
+            $url = env('APP_URL') . '/upcoming-orders/' . $order['orders_user_id'];
+        }
+
+//        checks for delivery or pickup
         if ($order['order_details']['method'] == 'Pickup') {
             $deliveryFee = 0;
         } else if ($order['order_details']['method'] == 'Delivery') {
             $deliveryFee = floatval($chef->delivery_fee);
         }
 
-        foreach($order['meal_details'] as $meal) {
-            $price = Meal::find($meal['meal_id'])->price;
-            $order_total = $order_total + ($price * $meal['quantity']);
+//        sets the bundle price if there is one
+        if ($order['bundle'] == true ) {
+            $order_total = $order['order_total'];
+        } else {
+            foreach ($order['meal_details'] as $meal) {
+                $price = Meal::find($meal['meal_id'])->price;
+                $order_total = $order_total + ($price * $meal['quantity']);
+            }
         }
 
         if ($order_total > 0) {
-            $order_total += ($deliveryFee + $serviceFee);
+            $order_total += ($deliveryFee);
         }
 
         $saved = $chef->orders()
             ->create(array(
-                'orders_user_id' => $order['orders_user_id'],
-                'orders_chef_id' => $chef->id,
+                'orders_user_id' => $order['orders_user_id'] ?? null,
+                'orders_chef_id' => $chef->chef_id,
                 'meal_details' => json_encode($order['meal_details']),
                 'customer_details' => json_encode($order['customer_details']),
                 'order_details' => json_encode($order['order_details']),
@@ -135,18 +167,30 @@ namespace OrchidEats\Http\Controllers;
             ));
 
         if ($saved) {
-            $user = User::find($order['orders_user_id']);
-            $user->cart()->delete();
-            $order['url'] = $url;
-            $order['buyer'] = $user->first_name;
-            \Mail::to($order['customer_details']['email'])->send(new NewOrder($order));
+            $user = User::find($order['orders_user_id']) ?? null;
+            if ($user != null) {
+                $user->cart()->delete();
+                $data['url'] = $url;
+                $data['buyer'] = $user->first_name;
+                \Mail::to($order['customer_details']['email'])->send(new NewOrder($data));
+            } else {
+                $data['order'] = $order['order_details'];
+                $data['meals'] = $order['meal_details'];
+                $data['total'] = $order_total;
+                $data['buyer'] = $order['customer_details']['name'];
+                $data['email'] = $email;
+                $data['phone'] = $phone;
+                \Mail::to($order['customer_details']['email'])->send(new NonUserOrder($data));
+            }
+
 
             return response()->json([
                 'status' => 'success'
             ]);
         } else {
             return response()->json([
-                'status' => 'error'
+                'status' => 'error',
+                'data' => [$email, $phone]
             ]);
         }
     }
